fix build issues in the VS. The command line build was not affected by this refactoring issues
